Set POV now before changing Pano
The major change is that the POV is now set to the POV of the next pano.
After that is done, then the pano is changed. This is to fix the error
of the camera always staying at the POV of the previous pano which
causes problem when doing things like traveling around a circle round
about.